fix: convert UserIdentity struct to map before structpb serialisation
structpb.NewValue only accepts primitives, maps, and slices — passing the
UserIdentity struct directly caused serialisation to fail. Convert to an
explicit map[string]interface{} before calling NewValue.
